More comprehensive fix for Issue #161

+++ LibreMetaverse/Primitives/TextureEntry.cs
@@ -978,29 +978,29 @@ namespace OpenMetaverse
                     {
                         #region Bitfield Setup
 
-                        uint[] textures = new uint[FaceTextures.Length];
+                        ulong[] textures = new ulong[FaceTextures.Length];
                         InitializeArray(ref textures);
-                        uint[] rgbas = new uint[FaceTextures.Length];
+                        ulong[] rgbas = new ulong[FaceTextures.Length];
                         InitializeArray(ref rgbas);
-                        uint[] repeatus = new uint[FaceTextures.Length];
+                        ulong[] repeatus = new ulong[FaceTextures.Length];
                         InitializeArray(ref repeatus);
-                        uint[] repeatvs = new uint[FaceTextures.Length];
+                        ulong[] repeatvs = new ulong[FaceTextures.Length];
                         InitializeArray(ref repeatvs);
-                        uint[] offsetus = new uint[FaceTextures.Length];
+                        ulong[] offsetus = new ulong[FaceTextures.Length];
                         InitializeArray(ref offsetus);
-                        uint[] offsetvs = new uint[FaceTextures.Length];
+                        ulong[] offsetvs = new ulong[FaceTextures.Length];
                         InitializeArray(ref offsetvs);
-                        uint[] rotations = new uint[FaceTextures.Length];
+                        ulong[] rotations = new ulong[FaceTextures.Length];
                         InitializeArray(ref rotations);
-                        uint[] materials = new uint[FaceTextures.Length];
+                        ulong[] materials = new ulong[FaceTextures.Length];
                         InitializeArray(ref materials);
-                        uint[] medias = new uint[FaceTextures.Length];
+                        ulong[] medias = new ulong[FaceTextures.Length];
                         InitializeArray(ref medias);
-                        uint[] glows = new uint[FaceTextures.Length];
+                        ulong[] glows = new ulong[FaceTextures.Length];
                         InitializeArray(ref glows);
-                        uint[] materialIDs = new uint[FaceTextures.Length];
+                        ulong[] materialIDs = new ulong[FaceTextures.Length];
                         InitializeArray(ref materialIDs);
-                        uint[] renderMaterialIDs = new uint[FaceTextures.Length];
+                        ulong[] renderMaterialIDs = new ulong[FaceTextures.Length];
                         InitializeArray(ref renderMaterialIDs);
 
                         for (int i = 0; i < FaceTextures.Length; i++)
@@ -1009,63 +1009,63 @@ namespace OpenMetaverse
 
                             if (FaceTextures[i].TextureID != DefaultTexture.TextureID)
                             {
-                                if (textures[i] == uint.MaxValue) textures[i] = 0;
-                                textures[i] |= (uint)(1 << i);
+                                if (textures[i] == ulong.MaxValue) textures[i] = 0;
+                                textures[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].RGBA != DefaultTexture.RGBA)
                             {
-                                if (rgbas[i] == uint.MaxValue) rgbas[i] = 0;
-                                rgbas[i] |= (uint)(1 << i);
+                                if (rgbas[i] == ulong.MaxValue) rgbas[i] = 0;
+                                rgbas[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].RepeatU != DefaultTexture.RepeatU)
                             {
-                                if (repeatus[i] == uint.MaxValue) repeatus[i] = 0;
-                                repeatus[i] |= (uint)(1 << i);
+                                if (repeatus[i] == ulong.MaxValue) repeatus[i] = 0;
+                                repeatus[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].RepeatV != DefaultTexture.RepeatV)
                             {
-                                if (repeatvs[i] == uint.MaxValue) repeatvs[i] = 0;
-                                repeatvs[i] |= (uint)(1 << i);
+                                if (repeatvs[i] == ulong.MaxValue) repeatvs[i] = 0;
+                                repeatvs[i] |= (1UL << i);
                             }
                             if (Helpers.TEOffsetShort(FaceTextures[i].OffsetU) != Helpers.TEOffsetShort(DefaultTexture.OffsetU))
                             {
-                                if (offsetus[i] == uint.MaxValue) offsetus[i] = 0;
-                                offsetus[i] |= (uint)(1 << i);
+                                if (offsetus[i] == ulong.MaxValue) offsetus[i] = 0;
+                                offsetus[i] |= (1UL << i);
                             }
                             if (Helpers.TEOffsetShort(FaceTextures[i].OffsetV) != Helpers.TEOffsetShort(DefaultTexture.OffsetV))
                             {
-                                if (offsetvs[i] == uint.MaxValue) offsetvs[i] = 0;
-                                offsetvs[i] |= (uint)(1 << i);
+                                if (offsetvs[i] == ulong.MaxValue) offsetvs[i] = 0;
+                                offsetvs[i] |= (1UL << i);
                             }
                             if (Helpers.TERotationShort(FaceTextures[i].Rotation) != Helpers.TERotationShort(DefaultTexture.Rotation))
                             {
-                                if (rotations[i] == uint.MaxValue) rotations[i] = 0;
-                                rotations[i] |= (uint)(1 << i);
+                                if (rotations[i] == ulong.MaxValue) rotations[i] = 0;
+                                rotations[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].material != DefaultTexture.material)
                             {
-                                if (materials[i] == uint.MaxValue) materials[i] = 0;
-                                materials[i] |= (uint)(1 << i);
+                                if (materials[i] == ulong.MaxValue) materials[i] = 0;
+                                materials[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].media != DefaultTexture.media)
                             {
-                                if (medias[i] == uint.MaxValue) medias[i] = 0;
-                                medias[i] |= (uint)(1 << i);
+                                if (medias[i] == ulong.MaxValue) medias[i] = 0;
+                                medias[i] |= (1UL << i);
                             }
                             if (Helpers.TEGlowByte(FaceTextures[i].Glow) != Helpers.TEGlowByte(DefaultTexture.Glow))
                             {
-                                if (glows[i] == uint.MaxValue) glows[i] = 0;
-                                glows[i] |= (uint)(1 << i);
+                                if (glows[i] == ulong.MaxValue) glows[i] = 0;
+                                glows[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].MaterialID != DefaultTexture.MaterialID)
                             {
-                                if (materialIDs[i] == uint.MaxValue) materialIDs[i] = 0;
-                                materialIDs[i] |= (uint)(1 << i);
+                                if (materialIDs[i] == ulong.MaxValue) materialIDs[i] = 0;
+                                materialIDs[i] |= (1UL << i);
                             }
                             if (FaceTextures[i].RenderMaterialID != DefaultTexture.RenderMaterialID)
                             {
-                                if (renderMaterialIDs[i] == uint.MaxValue) renderMaterialIDs[i] = 0;
-                                renderMaterialIDs[i] |= (uint)(1 << i);
+                                if (renderMaterialIDs[i] == ulong.MaxValue) renderMaterialIDs[i] = 0;
+                                renderMaterialIDs[i] |= (1UL << i);
                             }
                         }
 
@@ -1075,7 +1075,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.TextureID.GetBytes());
                         for (int i = 0; i < textures.Length; i++)
                         {
-                            if (textures[i] != uint.MaxValue)
+                            if (textures[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(textures[i]));
                                 binWriter.Write(FaceTextures[i].TextureID.GetBytes());
@@ -1089,7 +1089,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.RGBA.GetBytes(true));
                         for (int i = 0; i < rgbas.Length; i++)
                         {
-                            if (rgbas[i] != uint.MaxValue)
+                            if (rgbas[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(rgbas[i]));
                                 // Serialize the color bytes inverted to optimize for zerocoding
@@ -1103,7 +1103,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.RepeatU);
                         for (int i = 0; i < repeatus.Length; i++)
                         {
-                            if (repeatus[i] != uint.MaxValue)
+                            if (repeatus[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(repeatus[i]));
                                 binWriter.Write(FaceTextures[i].RepeatU);
@@ -1116,7 +1116,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.RepeatV);
                         for (int i = 0; i < repeatvs.Length; i++)
                         {
-                            if (repeatvs[i] != uint.MaxValue)
+                            if (repeatvs[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(repeatvs[i]));
                                 binWriter.Write(FaceTextures[i].RepeatV);
@@ -1129,7 +1129,7 @@ namespace OpenMetaverse
                         binWriter.Write(Helpers.TEOffsetShort(DefaultTexture.OffsetU));
                         for (int i = 0; i < offsetus.Length; i++)
                         {
-                            if (offsetus[i] != uint.MaxValue)
+                            if (offsetus[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(offsetus[i]));
                                 binWriter.Write(Helpers.TEOffsetShort(FaceTextures[i].OffsetU));
@@ -1142,7 +1142,7 @@ namespace OpenMetaverse
                         binWriter.Write(Helpers.TEOffsetShort(DefaultTexture.OffsetV));
                         for (int i = 0; i < offsetvs.Length; i++)
                         {
-                            if (offsetvs[i] != uint.MaxValue)
+                            if (offsetvs[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(offsetvs[i]));
                                 binWriter.Write(Helpers.TEOffsetShort(FaceTextures[i].OffsetV));
@@ -1155,7 +1155,7 @@ namespace OpenMetaverse
                         binWriter.Write(Helpers.TERotationShort(DefaultTexture.Rotation));
                         for (int i = 0; i < rotations.Length; i++)
                         {
-                            if (rotations[i] != uint.MaxValue)
+                            if (rotations[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(rotations[i]));
                                 binWriter.Write(Helpers.TERotationShort(FaceTextures[i].Rotation));
@@ -1168,7 +1168,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.material);
                         for (int i = 0; i < materials.Length; i++)
                         {
-                            if (materials[i] != uint.MaxValue)
+                            if (materials[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(materials[i]));
                                 binWriter.Write(FaceTextures[i].material);
@@ -1181,7 +1181,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.media);
                         for (int i = 0; i < medias.Length; i++)
                         {
-                            if (medias[i] != uint.MaxValue)
+                            if (medias[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(medias[i]));
                                 binWriter.Write(FaceTextures[i].media);
@@ -1194,7 +1194,7 @@ namespace OpenMetaverse
                         binWriter.Write(Helpers.TEGlowByte(DefaultTexture.Glow));
                         for (int i = 0; i < glows.Length; i++)
                         {
-                            if (glows[i] != uint.MaxValue)
+                            if (glows[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(glows[i]));
                                 binWriter.Write(Helpers.TEGlowByte(FaceTextures[i].Glow));
@@ -1207,7 +1207,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.MaterialID.GetBytes());
                         for (int i = 0; i < materialIDs.Length; i++)
                         {
-                            if (materialIDs[i] != uint.MaxValue)
+                            if (materialIDs[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(materialIDs[i]));
                                 binWriter.Write(FaceTextures[i].MaterialID.GetBytes());
@@ -1220,7 +1220,7 @@ namespace OpenMetaverse
                         binWriter.Write(DefaultTexture.RenderMaterialID.GetBytes());
                         for (int i = 0; i < renderMaterialIDs.Length; i++)
                         {
-                            if (renderMaterialIDs[i] != uint.MaxValue)
+                            if (renderMaterialIDs[i] != ulong.MaxValue)
                             {
                                 binWriter.Write(GetFaceBitfieldBytes(renderMaterialIDs[i]));
                                 binWriter.Write(FaceTextures[i].RenderMaterialID.GetBytes());
@@ -1261,10 +1261,10 @@ namespace OpenMetaverse
 
             #region Helpers
 
-            private void InitializeArray(ref uint[] array)
+            private void InitializeArray(ref ulong[] array)
             {
                 for (int i = 0; i < array.Length; i++)
-                    array[i] = uint.MaxValue;
+                    array[i] = ulong.MaxValue;
             }
 
             private bool ReadFaceBitfield(byte[] data, ref int pos, ref UInt64 faceBits, ref uint bitfieldSize)
@@ -1288,10 +1288,10 @@ namespace OpenMetaverse
                 return (faceBits != 0);
             }
 
-            private byte[] GetFaceBitfieldBytes(uint bitfield)
+            private byte[] GetFaceBitfieldBytes(ulong bitfield)
             {
                 int byteLength = 0;
-                uint tmpBitfield = bitfield;
+                ulong tmpBitfield = bitfield;
                 while (tmpBitfield != 0)
                 {
                     tmpBitfield >>= 7;
